+++ splink/internals/comparison_level.py
@@ -247,22 +247,40 @@ class ComparisonLevel:
         self._u_probability = value
 
     @property
-    def _m_probability_description(self):
+    def _m_probability_description(self) -> str:
+        if self.is_null_level:
+            return ""
         if self.m_probability is not None:
+            percentage = self.m_probability * 100
+            one_in_n = (
+                1 / self.m_probability if self.m_probability > 0 else float("inf")
+            )
             return (
                 "Amongst matching record comparisons, "
-                f"{self.m_probability:.2%} of records are in the "
+                f"{percentage:.4g}% of records (i.e. one in "
+                f"{self._num_fmt_dp_or_sf(one_in_n)}) are in the "
                 f"{self.label_for_charts.lower()} comparison level"
             )
+        else:
+            return ""
 
     @property
-    def _u_probability_description(self):
+    def _u_probability_description(self) -> str:
+        if self.is_null_level:
+            return ""
         if self.u_probability is not None:
+            percentage = self.u_probability * 100
+            one_in_n = (
+                1 / self.u_probability if self.u_probability > 0 else float("inf")
+            )
             return (
                 "Amongst non-matching record comparisons, "
-                f"{self.u_probability:.2%} of records are in the "
+                f"{percentage:.4g}% of records (i.e. one in "
+                f"{self._num_fmt_dp_or_sf(one_in_n)}) are in the "
                 f"{self.label_for_charts.lower()} comparison level"
             )
+        else:
+            return ""
 
     def _add_trained_u_probability(self, val, desc="no description given"):
         self._trained_u_probabilities.append(
@@ -352,21 +370,36 @@ class ComparisonLevel:
         else:
             return math.log2(self._bayes_factor)
 
+    def _num_fmt_dp_or_sf(self, val):
+        if val > 5000:
+            return f"{val:,.0f}"
+        elif val >= 100:
+            return f"{val:,.0f}"
+        else:
+            return f"{val:,.4g}"
+
     @property
     def _bayes_factor_description(self):
         text = (
             f"If comparison level is `{self.label_for_charts.lower()}` "
             "then comparison is"
         )
+
         if self._bayes_factor == math.inf:
             return f"{text} certain to be a match"
         elif self._bayes_factor == 0.0:
             return f"{text} impossible to be a match"
         elif self._bayes_factor >= 1.0:
-            return f"{text} {self._bayes_factor:,.2f} times more likely to be a match"
+            return (
+                f"{text} {self._num_fmt_dp_or_sf(self._bayes_factor)} times "
+                "more likely to be a match"
+            )
         else:
             mult = 1 / self._bayes_factor
-            return f"{text}  {mult:,.2f} times less likely to be a match"
+            return (
+                f"{text} {self._num_fmt_dp_or_sf(mult)} times "
+                "less likely to be a match"
+            )
 
     @property
     def label_for_charts(self):
@@ -679,6 +712,8 @@ class ComparisonLevel:
         output["comparison_vector_value"] = self.comparison_vector_value
         output["max_comparison_vector_value"] = comparison_num_levels - 1
         output["bayes_factor_description"] = self._bayes_factor_description
+        output["m_probability_description"] = self._m_probability_description
+        output["u_probability_description"] = self._u_probability_description
 
         return output